Create temporary Class PostProcessorReadyInterface for PostProcessor transition
